changes to security utils and register/search model group

+++ plugin/src/main/java/org/opensearch/ml/action/models/DeleteModelTransportAction.java
@@ -41,6 +41,7 @@ import org.opensearch.index.reindex.DeleteByQueryAction;
 import org.opensearch.index.reindex.DeleteByQueryRequest;
 import org.opensearch.ml.common.MLModel;
 import org.opensearch.ml.common.exception.MLResourceNotFoundException;
+import org.opensearch.ml.common.exception.MLValidationException;
 import org.opensearch.ml.common.model.MLModelState;
 import org.opensearch.ml.common.transport.model.MLModelDeleteAction;
 import org.opensearch.ml.common.transport.model.MLModelDeleteRequest;
@@ -107,44 +108,47 @@ public class DeleteModelTransportAction extends HandledTransportAction<ActionReq
                         MLModel mlModel = MLModel.parse(parser, algorithmName);
 
 
-                        if ((mlModel.getModelGroupId() != null)
-                            && (filterByEnabled)
-                            && (!SecurityUtils.validateModelGroupAccess(user, mlModel.getModelGroupId(), client))) {
-                            log.error("User doesn't have valid privilege to perform this operation");
-                            throw new IllegalArgumentException("User doesn't have valid privilege to perform this operation");
-                        }
+                        SecurityUtils.validateModelGroupAccess(user, mlModel.getModelGroupId(), client, ActionListener.wrap(access -> {
+                            if ((filterByEnabled) && (Boolean.FALSE.equals(access))) {
+                                actionListener
+                                    .onFailure(new MLValidationException("User Doesn't have previlege to perform this operation"));
+                            } else {
+                                MLModelState mlModelState = mlModel.getModelState();
+                                if (mlModelState.equals(MLModelState.LOADED)
+                                    || mlModelState.equals(MLModelState.LOADING)
+                                    || mlModelState.equals(MLModelState.PARTIALLY_LOADED)
+                                    || mlModelState.equals(MLModelState.DEPLOYED)
+                                    || mlModelState.equals(MLModelState.DEPLOYING)
+                                    || mlModelState.equals(MLModelState.PARTIALLY_DEPLOYED)) {
+                                    actionListener
+                                        .onFailure(
+                                            new Exception(
+                                                "Model cannot be deleted in deploying or deployed state. Try undeploy model first then delete"
+                                            )
+                                        );
+                                } else {
+                                    DeleteRequest deleteRequest = new DeleteRequest(ML_MODEL_INDEX, modelId);
+                                    client.delete(deleteRequest, new ActionListener<DeleteResponse>() {
+                                        @Override
+                                        public void onResponse(DeleteResponse deleteResponse) {
+                                            deleteModelChunks(modelId, deleteResponse, actionListener);
+                                        }
 
-                        MLModelState mlModelState = mlModel.getModelState();
-                        if (mlModelState.equals(MLModelState.LOADED)
-                            || mlModelState.equals(MLModelState.LOADING)
-                            || mlModelState.equals(MLModelState.PARTIALLY_LOADED)
-                            || mlModelState.equals(MLModelState.DEPLOYED)
-                            || mlModelState.equals(MLModelState.DEPLOYING)
-                            || mlModelState.equals(MLModelState.PARTIALLY_DEPLOYED)) {
-                            actionListener
-                                .onFailure(
-                                    new Exception(
-                                        "Model cannot be deleted in deploying or deployed state. Try undeploy model first then delete"
-                                    )
-                                );
-                        } else {
-                            DeleteRequest deleteRequest = new DeleteRequest(ML_MODEL_INDEX, modelId);
-                            client.delete(deleteRequest, new ActionListener<DeleteResponse>() {
-                                @Override
-                                public void onResponse(DeleteResponse deleteResponse) {
-                                    deleteModelChunks(modelId, deleteResponse, actionListener);
+                                        @Override
+                                        public void onFailure(Exception e) {
+                                            log.error("Failed to delete model meta data for model: " + modelId, e);
+                                            if (e instanceof ResourceNotFoundException) {
+                                                deleteModelChunks(modelId, null, actionListener);
+                                            }
+                                            actionListener.onFailure(e);
+                                        }
+                                    });
                                 }
-
-                                @Override
-                                public void onFailure(Exception e) {
-                                    log.error("Failed to delete model meta data for model: " + modelId, e);
-                                    if (e instanceof ResourceNotFoundException) {
-                                        deleteModelChunks(modelId, null, actionListener);
-                                    }
-                                    actionListener.onFailure(e);
-                                }
-                            });
-                        }
+                            }
+                        }, e -> {
+                            log.error("Failed to validate Access for Model Id " + modelId, e);
+                            actionListener.onFailure(e);
+                        }));
                     } catch (Exception e) {
                         log.error("Failed to parse ml model" + r.getId(), e);
                         actionListener.onFailure(e);
